Created util folder for scripts
