+++ scripts/utils/fetchUtils.ts
@@ -1,0 +1,31 @@
+
+
+// Parse the API tokens assuming they are in the environment variables
+const API_TOKENS = process.env.API_TOKENS;
+if (!API_TOKENS) {
+  throw new Error('API_TOKENS environment variable is not defined');
+}
+const tokens = API_TOKENS.split(',').reduce((acc, token) => {
+  const [name, value] = token.split(':');
+  acc[name] = value;
+  return acc;
+}, {} as Record<string, string>);
+
+
+// Function to fetch companies from a given API URL
+export async function fetchCompanies(baseURL: string | undefined) {
+  if (!baseURL) {
+    throw new Error(`No URL is provided.`)
+  }
+  const response = await fetch(`${baseURL}/companies`, {
+    method: 'GET',
+    headers: {
+      Authorization: `Bearer ${tokens['garbo']}`, // Use the appropriate token
+    },
+  });
+  if (!response.ok) {
+    throw new Error(`Failed to fetch data from ${baseURL}: ${response.statusText}`);
+  }
+  const data = await response.json();
+  return data;
+}
